fix(ci): rewrite smoke.mjs and ts-compat.mjs to simulate real consumers
smoke.mjs:
  - temp project installs packed tarballs via npm, then imports by
    package name (not raw .tgz path)
  - writes test.mjs/test.cjs files instead of fragile inlined eval
  - uses process.execPath to avoid PATH resolution issues asdf/shims
  - skips adapter-contract (bundles vitest — cannot import outside
    vitest context)

ts-compat.mjs:
  - temp project installs packed tarballs, then imports * as ns from
    each package by name (no .d.ts path imports, no export-parsing regex)
  - tests TS ~5.5.0 + latest × moduleResolution bundler + nodenext
  - excludes adapter-contract (generic Uint8Array<ArrayBuffer> — TS 5.7+)

quality.mjs: use process.execPath + dirname(execPath)/npx instead of
  relying on PATH to find npx (asdf shim layer broke child processes)

.size-limit.json → .size-limit.cjs (supports comments, CJS for tool compat)
fixtures.ts: cast Uint8Array literal to fix d.ts output (was emitting
  generic Uint8Array<ArrayBuffer>, broke TS <5.7)

AGENTS.md: add smoke/ts-compat to mandatory pre-commit list + CI gate
  warning for .github/ or scripts/ changes

+++ packages/adapter-contract/src/fixtures.ts
@@ -38,7 +38,9 @@ export const fixtures = {
       },
       {
         filename: "photo.png",
-        content: new Uint8Array([137, 80, 78, 71, 13, 10, 26, 10]),
+        content: new Uint8Array([
+          137, 80, 78, 71, 13, 10, 26, 10,
+        ]) as Uint8Array,
         contentType: "image/png",
       },
     ],
